+++ lib/helpers_plethora.dart
@@ -2,5 +2,10 @@ library helpers_plethora;
 
 export 'src/extensions/datetime_x.dart';
 export 'src/extensions/datetimerange_x.dart';
+export 'src/extensions/double_x.dart';
+export 'src/extensions/int_x.dart';
+export 'src/extensions/list_comparable_x.dart';
 export 'src/extensions/list_datetimerange_x.dart';
+export 'src/extensions/list_num_x.dart';
 export 'src/extensions/string_x.dart';
+export 'src/extensions/timeofday_x.dart';