+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## [0.2.0-nullsafety.1] - 2021-1-7.
+## [0.2.0-nullsafety.1] - 2021-1-9.
 
-* Added `lastMonthDay` as extension getter method for `DateTime`.
+* Added `lastMonthDay` as extension getter method for `DateTime`, and several other methods for representing the date-time as a string.
+* Added extension methods for `double` class.
+* Added extension methods for `int` class.
+* Added extension methods for `List<Comparable>` class.
+* Added extension methods for `List<num>` class.
 * Added extension methods for `String` class.
+* Added extension methods for `TimeOfDay` class.
 
 ## [0.1.0-nullsafety.1] - 2021-1-7.